Add files via upload
The primary purpose of the code provided is to set up a proxy Simple Mail Transfer Protocol (SMTP) server that can filter incoming emails based on a predetermined blocklist. This blocklist is defined in a text file, allowing for easy updates and adjustments.

The proxy server operates by listening to SMTP requests made to a specific port, authenticating these requests, then inspecting and parsing the incoming email data. If the email address, the domain of the email address, or the top-level domain (TLD) of the sender are found within the blocklist, the server refuses to forward the email, effectively blocking it. If none of these aspects are found in the blocklist, the server proceeds to forward the email to a predefined destination SMTP server. In this case, the destination is Mailtrap, an online service designed for testing email sending functionality in a safe and controlled environment.

A second piece of code, emailTest.js, serves the purpose of testing this system. It uses the Node.js library Nodemailer to send a test email to the proxy SMTP server. The test email's sender address, recipient, and content can be customized within this script.

This system is particularly useful in testing and development environments where there is a need to evaluate email functionality, specifically the ability to block certain emails based on a blocklist. By providing a layer of inspection and forwarding to the SMTP process, this server enhances the control developers have over incoming email traffic, offering the possibility to filter out undesirable content, potentially mitigating spam or harmful emails.